// app/types/network.ts
import { AxiosResponse } from 'axios'

export interface INetworkRequest {
  params?: any
  data?: any
  URL: string
  headers?: any
  isInMemory?: boolean
  id?: string
  auth?: any
}

export type AxiosReqMethod = (
  payload: INetworkRequest,
) => Promise<AxiosResponse>

export interface IRequestType {
  GET: AxiosReqMethod
  POST: AxiosReqMethod
  PUT: AxiosReqMethod
  DELETE: AxiosReqMethod
  PATCH: AxiosReqMethod
}

export enum StatusCode {
  OK = 200,
  CREATED = 201,
  NO_CONTENT = 204,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  METHOD_NOT_ALLOWED = 405,
  CONFLICT = 409,
  INTERNAL_SERVER_ERROR = 500,
  BAD_GATEWAY = 502,
  SERVICE_UNAVAILABLE = 503,
}